backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
awacs==2.5.0
aws-sam-translator==1.97.0
billiard==4.2.1
boto3==1.38.4
botocore==1.38.4
bracex==2.5.post1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
cfn-flip==1.3.0
cfn-lint==1.34.2
charset-normalizer==3.4.1
cli_exit_tools==1.2.7
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmudict==1.0.32
coloredlogs==15.0
cryptography==44.0.2
decouple==0.0.7
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-cors-headers==4.7.0
django-extensions==4.1
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
docker==7.1.0
drf-spectacular==0.28.0
elevenlabs==1.57.0
formic2==1.0.3
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
igittigitt==2.1.5
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflection==0.5.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kombu==5.5.3
lark==1.2.2
lib-detect-testenv==2.0.8
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.5
openai==1.76.0
packaging==25.0
pathspec==0.12.1
pgvector==0.4.1
pillow==11.2.1
pillow_heif==0.22.0
prompt_toolkit==3.0.51
psycopg==3.2.6
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyhcl==0.4.5
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-frontmatter==1.1.0
python-hcl2==7.2.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
replicate==1.0.6
requests==2.32.3
rpds-py==0.24.0
runway==2.8.2
s3transfer==0.12.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
sympy==1.14.0
textstat==0.7.5
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
troposphere==4.9.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcmatch==10.0
wcwidth==0.2.13
websockets==15.0.1
yamllint==1.37.0
zipp==3.21.0
